Adds uniqid, detects bug in button classname tracking
Previous update attempts to track array indexes by assigning the button classname to the array length at creation.
This fails above a single input because all list items are created at once so they all end up with the same className.

This update attempts to create a new state variable called pointer which will be updated to equal the array length and stored for later use.
The issue I am running into is that the pointer is consistently returning undefined, and I don't yet know why.